fix тип полей

+++ employees/models.py
@@ -269,7 +269,7 @@ class Department(models.Model):
         'users.DoctorProfile', on_delete=models.CASCADE, blank=True, null=True, verbose_name='Профиль пользователя, обновившего запись', related_name='employees_department_updated'
     )
     external_id = models.CharField(max_length=255, default=None, blank=True, null=True, help_text="Внешний ИД-код", db_index=True)
-    lunch_duration = models.IntegerField(default=None, blank=True, null=True, help_text="30, 60, 120")
+    lunch_duration = models.PositiveSmallIntegerField(default=None, blank=True, null=True, help_text="30, 60, 120")
 
     def __str__(self):
         return self.name
@@ -432,7 +432,7 @@ class EmployeePosition(models.Model):
     external_id = models.CharField(max_length=255, default=None, blank=True, null=True, help_text="Внешний ИД-код", db_index=True)
     date_employment = models.DateField(verbose_name="Дата приема на работу", help_text="2025-01-11", blank=True, null=True, default=None)
     date_dismissal = models.DateField(verbose_name="Дата увольнения", help_text="2025-02-01", blank=True, null=True, default=None)
-    lunch_duration = models.IntegerField(default=None, blank=True, null=True, help_text="30, 60, 120")
+    lunch_duration = models.PositiveSmallIntegerField(default=None, blank=True, null=True, help_text="30, 60, 120")
 
     def __str__(self):
         return f'{self.employee} — {self.position} (ставка {self.rate})'
